Create prepareRedirects function from createRedirects
